# Storing XML schema locally for improved performance