… scripts/xml-schema/docbook.xsd