Since the DocBook 5 schema is very unlikely to change, we can afford to store it locally. The performance differences are as follows:

When downloading:
* In the office: 1.2 seconds
* At home: 11 seconds

When stored locally:
* In the office: 1.2 seconds
* At home: 1.1 seconds

Disclaimer: My computer in the office is a different one from that at home.

Even if we need to change the schema for Pantheon to work, we can simply update and commit the XSDs in `$REPO_HOME/scripts/xml-schema` and keep on rolling.

Resolves #655.