@tradej 2017-09-21
@rhoads-zach, please test on Mac, thank you.